+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Major rework of native code to separate Cordova plugin layer from launch navigator functionality.
 - Remove obsolete Windows Phone 8 platform.
 - Add explicit `enableDebug()` API function to enable debug output.
+- Add plugin variable to enable specification of Google API key for geocoding on Android. Fixes [#211](https://github.com/dpa99c/phonegap-launch-navigator/issues/211).
 
 ## v4.2.1
 - Enable override of default OKHTTP version to prevent Gradle version collisions. Fixes [#193](https://github.com/dpa99c/phonegap-launch-navigator/issues/193).